% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_data}
\alias{pkg_data}
\title{Package Data File}
\usage{
pkg_data(...)
}
\arguments{
\item{...}{File name.}
}
\value{
A character vector of length one.
}
\description{
Retrieve the full path to a package data file.
}
\keyword{internal}
